[#40] 레이아웃 코드 SettingView 내부로 이동
